Move init() from display to main. Add include guards where necessary.
