+++ inc/microbit/microbitdisplay.h
@@ -1,4 +1,9 @@
 
+#ifndef __MICROPY_INCLUDED_MICROBIT_DISPLAY_H__
+#define __MICROPY_INCLUDED_MICROBIT_DISPLAY_H__
+
+#include "py/runtime.h"
+#include "microbitimage.h"
 
 typedef struct _microbit_display_obj_t {
     mp_obj_base_t base;
@@ -23,6 +28,9 @@ typedef struct _microbit_display_obj_t {
 
 extern microbit_display_obj_t microbit_display_obj;
 
+
+extern "C" {
+
 void microbit_display_print(microbit_display_obj_t *display, microbit_image_obj_t *image);
 
 void microbit_display_animate(microbit_display_obj_t *display, mp_obj_t iterable, mp_int_t delay, bool wait, bool loop);
@@ -33,5 +41,12 @@ mp_int_t microbit_display_get_pixel(microbit_display_obj_t *display, mp_int_t x,
 
 void microbit_display_set_pixel(microbit_display_obj_t *display, mp_int_t x, mp_int_t y, mp_int_t val);
 
-
 void microbit_display_clear(void);
+
+void microbit_display_init(void);
+
+void microbit_display_tick(void);
+
+}
+
+#endif // __MICROPY_INCLUDED_MICROBIT_DISPLAY_H__
